requirements.txt
python-dotenv

llama-index
llama-index-core
llama-index-readers-file
llama-index-llms-ollama
llama-index-llms-gemini
llama-index-embeddings-huggingface

trulens
trulens-apps-llamaindex
trulens-providers-openai
openai

trulens-eval
llama-index-llms-gemini
trulens-providers-openai
trulens-apps-llamaindex
llama_index.node_parser